# tak/docs/openapi.yaml
openapi: 3.0.0
info:
  title: TAK — TON Agent Kit API
  version: 0.1.0
  description: |
    Developer infrastructure for agent coordination on TON.

    **Key Principles:**
    - TAK coordinates. TON MCP executes. TAK never holds funds.
    - All operations are idempotent via `idempotency_key`
    - Deals are immutable once created
    - Approval required before execution
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.tak.ton
    description: Production server
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error_code:
          type: string
          example: NOT_FOUND
        message:
          type: string
        trace_id:
          type: string
          example: trace_1a2b3c4d5e
...
  responses:
    Error:
      description: General error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
paths:
...
  /api/services:
    get:
      summary: List all services
      tags: [Services]
      responses:
        "200":
          description: OK
    post:
      summary: Publish service
      tags: [Services]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [agent_id, service_name, base_price_nano]
              properties:
                agent_id:
                  type: string
                service_name:
                  type: string
                base_price_nano:
                  type: integer
      responses:
        "201":
          description: Created
  /api/services/search:
    get:
      summary: Search services
      tags: [Services]
      parameters:
        - in: query
          name: tag
          schema:
            type: string
        - in: query
          name: max_price
          schema:
            type: integer
      responses:
        "200":
          description: Found services
  /api/services/{id}:
    get:
      summary: Get service
      tags: [Services]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    put:
      summary: Update service
      tags: [Services]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Updated
    delete:
      summary: Delete service
      tags: [Services]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
...
  /api/deals:
    get:
      summary: List deals
      tags: [Deals]
      responses:
        "200":
          description: OK
    post:
      summary: Create deal from offer
      tags: [Deals]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [request_id, offer_id]
              properties:
                request_id:
                  type: string
                offer_id:
                  type: string
                execution_type:
                  type: string
                  enum: [ton_transfer, jetton_transfer, nft_transfer, contract_call]
                execution_payload:
                  type: object
      responses:
        "201":
          description: Created
  /api/deals/{id}:
    get:
      summary: Get deal
      tags: [Deals]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    put:
      summary: Cancel deal
      tags: [Deals]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [cancelled]
      responses:
        "200":
          description: Cancelled
    delete:
      summary: Delete deal
      tags: [Deals]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
  /api/deals/{id}/approve:
    post:
      summary: Approve deal
      tags: [Deals]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Approved
  /api/deals/{id}/execute:
    post:
      summary: Execute deal via MCPAdapter
      tags: [Deals]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Executed
        "500":
          $ref: '#/components/responses/Error'
  /api/deals/{id}/reject:
    post:
      summary: Reject deal
      tags: [Deals]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Rejected
...